Move custom ui logic to a separate module
This will enable people to choose if they want to pull in the extra
dependencies or not. Ultimately making their apps smaller.
