+++ chromecast-sender-sample-app/src/main/res/drawable/ayp_ic_pause_36dp.xml
