+++ chromecast-sender-sample-app/src/main/res/drawable/ayp_ic_play_36dp.xml
