+++ chromecast-sender-sample-app/src/main/res/drawable/ayp_ic_youtube_24dp.xml
