+++ core-custom-ui/build.gradle
@@ -13,24 +13,24 @@ android {
   sourceSets {
     main.res.srcDirs = [
             'src/main/res',
+            'src/main/res-public'
     ]
   }
-  namespace 'com.pierfrancescosoffritti.androidyoutubeplayer'
+
+  namespace 'com.pierfrancescosoffritti.androidyoutubeplayer.core.customui'
 }
 
 dependencies {
-  testImplementation "junit:junit:$versions.junit"
-  androidTestImplementation "androidx.test:runner:$versions.runner"
-  androidTestImplementation "androidx.test.espresso:espresso-core:$versions.espressoCore"
-
-  api "androidx.lifecycle:lifecycle-runtime-ktx:$versions.androidxLifecycleRuntime"
+  api project(':core')
+  implementation "androidx.core:core-ktx:$versions.androidxCore"
+  implementation "androidx.recyclerview:recyclerview:$versions.androidxRecyclerView"
 }
 
 ext {
   PUBLISH_GROUP_ID = 'com.pierfrancescosoffritti.androidyoutubeplayer'
   PUBLISH_VERSION = versions.publishVersion_core
-  PUBLISH_ARTIFACT_ID = 'core'
-  PUBLISH_DESCRIPTION = 'YouTube Player library for Android, stable and customizable.'
+  PUBLISH_ARTIFACT_ID = 'core-custom-ui'
+  PUBLISH_DESCRIPTION = 'Custom UI to use with android-youtube-player core module.'
 }
 
 apply from: "../scripts/publish-module.gradle"
