+++ core-custom-ui/src/main/res-public/values/public.xml
