+++ core-custom-ui/src/main/res/drawable-v21/ayp_background_item_selected.xml
