+++ core-custom-ui/src/main/res/drawable/ayp_background_item_selected.xml
