+++ core-custom-ui/src/main/res/drawable/ayp_drop_shadow_top.xml
