+++ core-custom-ui/src/main/res/drawable/ayp_ic_fullscreen_24dp.xml
