+++ core-custom-ui/src/main/res/drawable/ayp_ic_fullscreen_exit_24dp.xml
