+++ core-custom-ui/src/main/res/drawable/ayp_ic_menu_24dp.xml
