+++ core-custom-ui/src/main/res/drawable/ayp_shape_rounded_corners.xml
