+++ core-custom-ui/src/main/res/layout/ayp_default_player_ui.xml
@@ -107,7 +107,7 @@
             android:layout_alignParentStart="true"
             android:layout_alignParentEnd="true" >
 
-            <com.pierfrancescosoffritti.androidyoutubeplayer.core.ui.views.YouTubePlayerSeekBar
+            <com.pierfrancescosoffritti.androidyoutubeplayer.core.customui.views.YouTubePlayerSeekBar
                 android:id="@+id/youtube_player_seekbar"
                 android:layout_width="0dp"
                 android:layout_height="wrap_content"
