+++ core-custom-ui/src/main/res/layout/ayp_menu_item.xml
