+++ core-custom-ui/src/main/res/layout/ayp_player_menu.xml
