+++ core-custom-ui/src/main/res/values/colors.xml
