+++ core-custom-ui/src/main/res/values/dimens.xml
