+++ core-custom-ui/src/main/res/values/strings.xml
