+++ core-sample-app/build.gradle
@@ -40,9 +40,10 @@ dependencies {
   androidTestImplementation "androidx.test.espresso:espresso-core:$versions.espressoCore"
 
   implementation project(':core')
+  implementation project(':core-custom-ui')
   implementation project(':chromecast-sender')
 
-  implementation "androidx.appcompat:appcompat:$versions.androidxCore"
+  implementation "androidx.appcompat:appcompat:$versions.appcompat"
   implementation "androidx.recyclerview:recyclerview:$versions.androidxRecyclerView"
   implementation "androidx.constraintlayout:constraintlayout:$versions.androidxConstraintLayout"
