+++ core/src/main/res/values/attrs.xml
@@ -1,10 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <declare-styleable name="YouTubePlayerSeekBar">
-        <attr name="fontSize" format="dimension" />
-        <attr name="color" format="color" />
-    </declare-styleable>
-
     <declare-styleable name="YouTubePlayerView">
         <attr name="enableAutomaticInitialization" format="boolean" />
         <attr name="videoId" format="string" />
