+++ dependencies.gradle
@@ -19,7 +19,8 @@ ext.versions = [
         kotlin                          : '1.8.0',
 
         // AndroidX
-        androidxCore                    : '1.6.1',
+        androidxCore                    : '1.9.0',
+        appcompat                       : '1.6.1',
         androidxAnnotations             : '1.6.0',
         androidxConstraintLayout        : '2.1.4',
         androidxRecyclerView            : '1.3.0',
