+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':core', ':core-sample-app', ':chromecast-sender', ':chromecast-sender-sample-app'
+include ':core', ':core-sample-app', ':chromecast-sender', ':chromecast-sender-sample-app', ':core-custom-ui'
